# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AgeEstimationRequest": {
        "additionalProperties": false,
        "properties": {
          "task": {
            "const": "age_estimation",
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task"
        ],
        "title": "AgeEstimationRequest",
        "type": "object"
      },
      "AgeEstimationResult": {
        "additionalProperties": false,
        "properties": {
          "disclaimer_code": {
            "const": "age_estimate_v1",
            "default": "age_estimate_v1",
            "title": "Disclaimer Code",
            "type": "string"
          },
          "estimate_years": {
            "title": "Estimate Years",
            "type": "integer"
          },
          "interval_coverage": {
            "description": "Nominal coverage of range_years, e.g. 0.8.",
            "title": "Interval Coverage",
            "type": "number"
          },
          "range_years": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Range Years",
            "type": "array"
          }
        },
        "required": [
          "estimate_years",
          "range_years",
          "interval_coverage"
        ],
        "title": "AgeEstimationResult",
        "type": "object"
      },
      "AgeTransformationParams": {
        "additionalProperties": false,
        "properties": {
          "target_age_group": {
            "$ref": "#/components/schemas/TargetAgeGroup"
          }
        },
        "required": [
          "target_age_group"
        ],
        "title": "AgeTransformationParams",
        "type": "object"
      },
      "AgeTransformationRequest": {
        "additionalProperties": false,
        "properties": {
          "params": {
            "$ref": "#/components/schemas/AgeTransformationParams"
          },
          "task": {
            "const": "age_transformation",
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task",
          "params"
        ],
        "title": "AgeTransformationRequest",
        "type": "object"
      },
      "AgeTransformationResult": {
        "additionalProperties": false,
        "properties": {
          "disclaimer_code": {
            "const": "synthetic_image_v1",
            "default": "synthetic_image_v1",
            "title": "Disclaimer Code",
            "type": "string"
          },
          "image_url": {
            "title": "Image Url",
            "type": "string"
          },
          "image_url_expires_at": {
            "format": "date-time",
            "title": "Image Url Expires At",
            "type": "string"
          },
          "synthetic": {
            "const": true,
            "default": true,
            "title": "Synthetic",
            "type": "boolean"
          },
          "target_age_group": {
            "$ref": "#/components/schemas/TargetAgeGroup"
          },
          "watermarked": {
            "const": true,
            "default": true,
            "title": "Watermarked",
            "type": "boolean"
          }
        },
        "required": [
          "target_age_group",
          "image_url",
          "image_url_expires_at"
        ],
        "title": "AgeTransformationResult",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "enum": [
                  "not_found",
                  "method_not_allowed",
                  "http_error"
                ],
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "invalid_file_type",
          "file_too_large",
          "image_too_small",
          "image_too_large",
          "corrupt_image",
          "consent_required",
          "no_face_detected",
          "multiple_faces_detected",
          "face_too_small",
          "unsupported_target_age_group",
          "feature_disabled",
          "rate_limited",
          "session_not_found",
          "job_not_found",
          "result_not_found",
          "job_not_cancellable",
          "adults_only",
          "model_error",
          "inference_timeout",
          "validation_error",
          "internal_error"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "JobError",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobView": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Params",
            "type": "object"
          },
          "progress": {
            "title": "Progress",
            "type": "number"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgeEstimationResult"
              },
              {
                "$ref": "#/components/schemas/PresentationEstimationResult"
              },
              {
                "$ref": "#/components/schemas/AgeTransformationResult"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "task": {
            "$ref": "#/components/schemas/Task"
          }
        },
        "required": [
          "job_id",
          "session_id",
          "task",
          "params",
          "status",
          "progress",
          "stage",
          "created_at",
          "finished_at"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "ModelRef": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_mock": {
            "title": "Is Mock",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "version",
          "is_mock"
        ],
        "title": "ModelRef",
        "type": "object"
      },
      "PresentationEstimationRequest": {
        "additionalProperties": false,
        "properties": {
          "task": {
            "const": "presentation_estimation",
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task"
        ],
        "title": "PresentationEstimationRequest",
        "type": "object"
      },
      "PresentationEstimationResult": {
        "additionalProperties": false,
        "properties": {
          "disclaimer_code": {
            "const": "presentation_estimate_v1",
            "default": "presentation_estimate_v1",
            "title": "Disclaimer Code",
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/PresentationOutcome"
          },
          "scores": {
            "$ref": "#/components/schemas/PresentationScores"
          },
          "uncertain_threshold": {
            "title": "Uncertain Threshold",
            "type": "number"
          }
        },
        "required": [
          "outcome",
          "scores",
          "uncertain_threshold"
        ],
        "title": "PresentationEstimationResult",
        "type": "object"
      },
      "PresentationOutcome": {
        "enum": [
          "feminine_presenting",
          "masculine_presenting",
          "uncertain"
        ],
        "title": "PresentationOutcome",
        "type": "string"
      },
      "PresentationScores": {
        "additionalProperties": false,
        "properties": {
          "feminine_presenting": {
            "title": "Feminine Presenting",
            "type": "number"
          },
          "masculine_presenting": {
            "title": "Masculine Presenting",
            "type": "number"
          }
        },
        "required": [
          "feminine_presenting",
          "masculine_presenting"
        ],
        "title": "PresentationScores",
        "type": "object"
      },
...
      "TargetAgeGroup": {
        "description": "Creative targets for the illustrative aging tool. Not predictions.",
        "enum": [
          "child",
          "teen",
          "young_adult",
          "middle_aged_adult",
          "older_adult"
        ],
        "title": "TargetAgeGroup",
        "type": "string"
      },
...
      "Task": {
        "enum": [
          "age_estimation",
          "presentation_estimation",
          "age_transformation"
        ],
        "title": "Task",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Illustrative, uncertain estimates from a face image. Not for identification or any consequential decision. See docs/PLAN.md.",
    "title": "FaceLens API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs/{job_id}": {
      "delete": {
        "description": "Best-effort cancel. A running model call finishes, but its output is discarded.",
        "operationId": "cancel_job_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/results/{job_id}/image": {
      "get": {
        "description": "Short-lived signed URL, so the URL itself is the credential and <img> tags work.",
        "operationId": "get_result_image_v1_results__job_id__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exp",
            "required": true,
            "schema": {
              "title": "Exp",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sig",
            "required": true,
            "schema": {
              "maxLength": 128,
              "title": "Sig",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {}
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Result Image",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/sessions/{session_id}/jobs": {
      "post": {
        "operationId": "create_job_v1_sessions__session_id__jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AgeEstimationRequest"
                  },
                  {
                    "$ref": "#/components/schemas/PresentationEstimationRequest"
                  },
                  {
                    "$ref": "#/components/schemas/AgeTransformationRequest"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    }
  }
}